+++ src/main/java/janggi/controller/JanggiController.java
@@ -3,9 +3,8 @@ package janggi.controller;
 import janggi.domain.Board;
 import janggi.domain.GameInitializeMode;
 import janggi.domain.Team;
-import janggi.domain.dto.GameSession;
+import janggi.domain.GameSession;
 import janggi.domain.Position;
-import janggi.domain.strategy.BasicPlacementStrategy;
 import janggi.service.JanggiService;
 import janggi.util.InputParser;
 import janggi.view.InputView;
