+++ src/main/java/janggi/dao/PieceEntity.java
@@ -1,4 +1,4 @@
-package janggi.domain.dto;
+package janggi.dao;
 
 import janggi.domain.Team;
 import janggi.domain.piece.PieceType;
